fix windows build.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6218 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_channel.c
@@ -1421,11 +1421,11 @@ SWITCH_DECLARE(char *) switch_channel_expand_variables(switch_channel_t *channel
 
 					if (offset >= 0) {
 						sub_val += offset;
-					} else if (abs(offset) <= strlen(sub_val)) {
+					} else if ((size_t)abs(offset) <= strlen(sub_val)) {
 						sub_val = cloned_sub_val + (strlen(cloned_sub_val) + offset);
 					}
 
-					if (ooffset > 0 && ooffset < strlen(sub_val)) {
+					if (ooffset > 0 && (size_t)ooffset < strlen(sub_val)) {
 						if ((p = (char *)sub_val + ooffset)) {
 							*p = '\0';
 						}
